Fix some compiler warnings

+++ src/base/global.cpp
@@ -68,7 +68,7 @@ void writelogendl()
     app()->writelogendl();
 }
 
-void writeline(char repeat, int count, bool endl_after, bool endl_before)
+void writeline(char repeat, size_t count, bool endl_after, bool endl_before)
 {
     if (endl_before) {
         writelogendl();
